backend/app/requirements.txt
fastapi==0.99.0
pymysql==1.1.0
python-dotenv==1.0.1
uvicorn[standard]==0.22.0
python-multipart==0.0.9
PyJWT==2.9.0
passlib==1.7.4
python-jose==3.3.0
email-validator==2.1.0
cryptography==42.0.0
bcrypt==4.1.2
pydantic==1.10.12
pydantic[email] 

